backend/requirements.txt
alembic==0.9.9
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.2
psycopg2==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
stripe==1.80.0
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
XlsxWriter==1.0.2
